Create a Query base class and shared with XQuery and NQuery
- rename N1QLQuery to NQuery to be consistent with XQuery
- Keep only json related query functions in XQuery class
