+++ src/Couchbase.Lite.Shared/Couchbase.Lite.Shared.projitems
@@ -115,6 +115,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)Document\MRoot.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\IndexConfiguration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Query\N1QLQuery.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\NQuery.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Query\QueryBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Serialization\FleeceMutableArray.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Linq\IDocumentModel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Linq\ISelectResultContainer.cs" />
