+++ src/Couchbase.Lite.Shared/Query/XQuery.cs
@@ -19,16 +19,11 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
 
 using Couchbase.Lite.Internal.Logging;
 using Couchbase.Lite.Query;
 using Couchbase.Lite.Util;
-
-using JetBrains.Annotations;
 
 using LiteCore.Interop;
 
@@ -36,37 +31,13 @@ using Newtonsoft.Json;
 
 namespace Couchbase.Lite.Internal.Query
 {
-    internal class XQuery : IQuery, IStoppable
+    internal class XQuery : QueryBase
     {
-        #region Constants
-
-        private const string Tag = nameof(XQuery);
-        private static readonly TimeSpan DefaultLiveQueryUpdateInterval = TimeSpan.FromMilliseconds(200);
-
-        #endregion
-
         #region Variables
-
-        [NotNull]private readonly Event<QueryChangedEventArgs> _changed = new Event<QueryChangedEventArgs>();
-        [NotNull] internal readonly DisposalWatchdog _disposalWatchdog = new DisposalWatchdog(nameof(IQuery));
-        private readonly TimeSpan _updateInterval;
-
-        internal unsafe C4Query* _c4Query;
-        internal Dictionary<string, int> _columnNames;
-        private ListenerToken _databaseChangedToken;
-        [NotNull]internal List<QueryResultSet> _history = new List<QueryResultSet>();
-        private DateTime _lastUpdatedAt;
-        private int _observingCount = 0;
-        [NotNull]private Parameters _queryParameters = new Parameters();
-        private AtomicBool _willUpdate = false;
-        private bool _updating = false, _stopping = false;
-
+        private const string Tag = nameof(XQuery);
         #endregion
 
         #region Properties
-
-        public Database Database { get; set; }
-
         protected bool Distinct { get; set; }
 
         protected QueryDataSource FromImpl { get; set; }
@@ -87,27 +58,58 @@ namespace Couchbase.Lite.Internal.Query
 
         protected QueryExpression WhereImpl { get; set; }
 
-        public Parameters Parameters
-        {
-            get => _queryParameters;
-            set {
-                _queryParameters = value.Freeze();
-                Update();
-            }
-        }
-
         #endregion
 
-        #region Constructors
-
-        public XQuery()
+        #region Public Methods
+        public override unsafe IResultSet Execute()
         {
-            _updateInterval = DefaultLiveQueryUpdateInterval;
+            if (Database == null) {
+                throw new InvalidOperationException(CouchbaseLiteErrorMessage.InvalidQueryDBNull);
+            }
+
+            var fromImpl = FromImpl;
+            if (SelectImpl == null || fromImpl == null) {
+                throw new InvalidOperationException(CouchbaseLiteErrorMessage.InvalidQueryMissingSelectOrFrom);
+            }
+
+            var options = C4QueryOptions.Default;
+            var paramJson = Parameters.FLEncode();
+
+            var e = (C4QueryEnumerator*)fromImpl.ThreadSafety.DoLockedBridge(err =>
+            {
+                if (_disposalWatchdog.IsDisposed) {
+                    return null;
+                }
+
+                if (_c4Query == null) {
+                    Check();
+                }
+
+                var localOpts = options;
+                return NativeRaw.c4query_run(_c4Query, &localOpts, (FLSlice)paramJson, err);
+            });
+
+            paramJson.Dispose();
+
+            if (e == null) {
+                return new NullResultSet();
+            }
+
+            var retVal = new QueryResultSet(this, fromImpl.ThreadSafety, e, _columnNames);
+            _history.Add(retVal);
+            return retVal;
         }
 
-        ~XQuery()
+        public override unsafe string Explain()
         {
-            Dispose(true);
+            _disposalWatchdog.CheckDisposed();
+
+            // Used for debugging
+            if (_c4Query == null) {
+                Check();
+            }
+
+            return FromImpl?.ThreadSafety?.DoLocked(() => Native.c4query_explain(_c4Query)) ?? "(Unable to explain)";
         }
 
         #endregion
@@ -129,13 +131,13 @@ namespace Couchbase.Lite.Internal.Query
             SkipValue = source.SkipValue;
         }
 
-        protected virtual unsafe void Dispose(bool finalizing)
+        protected override unsafe void Dispose(bool finalizing)
         {
             if (!finalizing) {
                 Stop();
                 FromImpl.ThreadSafety.DoLocked(() =>
                 {
-                    foreach (var e in _history) {
+                    foreach (var e in _history)  {
                         e.Release();
                     }
 
@@ -144,8 +146,7 @@ namespace Couchbase.Lite.Internal.Query
                     _c4Query = null;
                     _disposalWatchdog.Dispose();
                 });
-            }
-            else {
+            } else {
                 // Database is not valid inside finalizer, but thread safety
                 // is guaranteed
                 Native.c4query_release(_c4Query);
@@ -153,7 +154,7 @@ namespace Couchbase.Lite.Internal.Query
             }
         }
 
-        protected void ValidateParams<T>(T[] param, [CallerMemberName]string tag = null)
+        protected void ValidateParams<T>(T[] param, [CallerMemberName] string tag = null)
         {
             if (param.Length == 0) {
                 var message = String.Format(CouchbaseLiteErrorMessage.ExpressionsMustContainOnePlusElement, tag);
@@ -162,15 +163,6 @@ namespace Couchbase.Lite.Internal.Query
         }
 
         #endregion
-
-        public void Stop()
-        {
-            if (_updating) {
-                _stopping = true;
-            } else {
-                Stopped();
-            }
-        }
 
         #region Private Methods
 
@@ -190,7 +182,7 @@ namespace Couchbase.Lite.Internal.Query
                 }
 
                 var query = Native.c4query_new2(Database.c4db, C4QueryLanguage.JSONQuery, jsonData, null, err);
-                if(query == null) {
+                if (query == null) {
                     return false;
                 }
 
@@ -220,12 +212,13 @@ namespace Couchbase.Lite.Internal.Query
                 }
 
                 if (map.ContainsKey(titleStr)) {
-                    throw new CouchbaseLiteException(C4ErrorCode.InvalidQuery, 
+                    throw new CouchbaseLiteException(C4ErrorCode.InvalidQuery,
                         String.Format(CouchbaseLiteErrorMessage.DuplicateSelectResultName, titleStr));
                 }
 
                 map.Add(titleStr, i);
             }
+
             return map;
         }
 
@@ -271,215 +264,21 @@ namespace Couchbase.Lite.Internal.Query
                 parameters["FROM"] = joinJson;
             } else {
                 var fromJson = FromImpl?.ToJSON();
-                if(fromJson != null) {
+                if (fromJson != null) {
                     parameters["FROM"] = new[] { fromJson };
                 }
             }
-            
+
             if (GroupByImpl != null) {
                 parameters["GROUP_BY"] = GroupByImpl.ToJSON();
             }
-            
+
             if (HavingImpl != null) {
                 parameters["HAVING"] = HavingImpl.ToJSON();
             }
 
             return JsonConvert.SerializeObject(parameters);
         }
-
-        private void OnDatabaseChanged(object sender, DatabaseChangedEventArgs e)
-        {
-            if (_willUpdate) {
-                return;
-            }
-
-            // External updates should poll less frequently
-            var updateInterval = _updateInterval;
-
-            var updateDelay = _lastUpdatedAt + updateInterval - DateTime.Now;
-            UpdateAfter(updateDelay);
-        }
-
-        private void Stopped()
-        {
-            Database?.RemoveActiveStoppable(this);
-            Database?.RemoveChangeListener(_databaseChangedToken);
-            _stopping = false;
-        }
-
-        private void Update()
-        {
-            if (!_willUpdate)
-                return;
-
-            _updating = true;
-
-            WriteLog.To.Query.I(Tag, $"{this}: Querying...");
-            var oldEnum = _history.LastOrDefault();
-            QueryResultSet newEnum = null;
-            Exception error = null;
-            if (oldEnum == null) {
-                try {
-                    var result = Execute();
-                    if (result is NullResultSet) {
-                        return;
-                    }
-
-                    newEnum = (QueryResultSet) result;
-                } catch (Exception e) {
-                    error = e;
-                }
-            } else {
-                newEnum = oldEnum.Refresh();
-                if (newEnum != null) {
-                    _history.Add(newEnum);
-                }
-            }
-
-            _updating = false;
-            _willUpdate.Set(false);
-
-            if (_stopping) {
-                Stopped();
-                return;
-            }
-
-            _lastUpdatedAt = DateTime.Now;
-
-            var changed = true;
-            if (newEnum != null) {
-                if (oldEnum != null) {
-                    WriteLog.To.Query.I(Tag, $"{this}: Changed!");
-                }
-            } else if (error != null) {
-                WriteLog.To.Query.E(Tag, $"{this}: Update failed: {error}");
-            } else {
-                changed = false;
-                WriteLog.To.Query.V(Tag, $"{this}: ...no change");
-            }
-
-            if (changed) {
-                _changed.Fire(this, new QueryChangedEventArgs(newEnum, error));
-            }
-        }
-
-        private async void UpdateAfter(TimeSpan updateDelay)
-        {
-            if (_willUpdate.Set(true)) {
-                return;
-            }
-
-            if (updateDelay > TimeSpan.Zero) {
-                await Task.Delay(updateDelay).ConfigureAwait(false);
-            }
-            
-            if (_willUpdate) {
-                Update();
-            }
-        }
-
-        #endregion
-
-        #region IDisposable
-
-        public void Dispose()
-        {
-            Dispose(false);
-        }
-
-        #endregion
-
-        #region IQuery
-
-        public ListenerToken AddChangeListener(TaskScheduler scheduler, [NotNull]EventHandler<QueryChangedEventArgs> handler)
-        {
-            CBDebug.MustNotBeNull(WriteLog.To.Query, Tag, nameof(handler), handler);
-            _disposalWatchdog.CheckDisposed();
-
-            var cbHandler = new CouchbaseEventHandler<QueryChangedEventArgs>(handler, scheduler);
-            _changed.Add(cbHandler);
-
-            if (Interlocked.Increment(ref _observingCount) == 1) {
-                Database?.AddActiveStoppable(this);
-                if (Database != null) {
-                    _databaseChangedToken = Database.AddChangeListener(OnDatabaseChanged);
-                } else {
-                    WriteLog.To.Query.W(Tag, @"Attempting to add a change listener onto a query with a null Database.  
-                                        Changed events will not continue to fire");
-                }
-
-                UpdateAfter(new TimeSpan(0));
-            }
-
-            return new ListenerToken(cbHandler, "query");
-        }
-
-        public ListenerToken AddChangeListener([NotNull]EventHandler<QueryChangedEventArgs> handler)
-        {
-            return AddChangeListener(null, handler);
-        }
-
-        public virtual unsafe IResultSet Execute()
-        {
-            if (Database == null) {
-                throw new InvalidOperationException(CouchbaseLiteErrorMessage.InvalidQueryDBNull);
-            }
-
-            var fromImpl = FromImpl;
-            if (SelectImpl == null || fromImpl == null) {
-                throw new InvalidOperationException(CouchbaseLiteErrorMessage.InvalidQueryMissingSelectOrFrom);
-            }
-            
-            var options = C4QueryOptions.Default;
-            var paramJson = Parameters.FLEncode();
-
-            var e = (C4QueryEnumerator*) fromImpl.ThreadSafety.DoLockedBridge(err =>
-            {
-                if (_disposalWatchdog.IsDisposed) {
-                    return null;
-                }
-
-                if (_c4Query == null) {
-                    Check();
-                }
-
-                var localOpts = options;
-                return NativeRaw.c4query_run(_c4Query, &localOpts, (FLSlice)paramJson, err);
-            });
-
-            paramJson.Dispose();
-
-            if (e == null) {
-                return new NullResultSet();
-            }
-
-            var retVal = new QueryResultSet(this, fromImpl.ThreadSafety, e, _columnNames);
-            _history.Add(retVal);
-            return retVal;
-        }
-        public unsafe string Explain()
-        {
-            _disposalWatchdog.CheckDisposed();
-
-            // Used for debugging
-            if (_c4Query == null) {
-                Check();
-            }
-
-            return FromImpl?.ThreadSafety?.DoLocked(() => Native.c4query_explain(_c4Query)) ?? "(Unable to explain)";
-        }
-
-        public void RemoveChangeListener(ListenerToken token)
-        {
-            _disposalWatchdog.CheckDisposed();
-            _changed.Remove(token);
-            if (Interlocked.Decrement(ref _observingCount) == 0) {
-                Stop();
-            }
-
-            _willUpdate.Set(false);
-        }
-
         #endregion
     }
 }
